requirements.txt
absl-py==1.3.0
accelerate==0.20.3
brotlipy==0.7.0
certifi==2023.11.17
cffi==1.15.1
cmake==3.27.7
cryptography==38.0.1
debugpy==1.5.1
easydict==1.11
#faiss==1.7.2
filelock==3.13.1
gmpy2==2.1.2
google-auth-oauthlib==0.4.6
idna==3.4
imageio==2.33.0
jinja2==3.1.2
joblib==1.3.2
jupyter-core==4.12.0
kiwisolver==1.4.2
lazy_loader==0.3
lit==17.0.6
markdown==3.5.1
markupsafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
mxnet==1.6.0
networkx==3.1
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.8.1.78
pillow==9.2.0
protobuf==3.20.3
psutil==5.9.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygments==2.15.1
pyqt5-sip==12.11.0
pysocks==1.7.1
pyyaml==6.0.1
pyzmq==25.1.0
requests==2.28.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.8.1
setuptools==65.5.0
sip==6.6.2
sympy==1.11.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tornado==6.1
triton==2.0.0
typing-extensions==4.3.0
urllib3==1.26.12
wheel==0.41.2
timm==0.4.12
py-blueprint==0.0.1.post3
mmsegmentation==0.18.0
ftfy==6.0.1
einops==0.3.2
pandas==2.2.0
medpy==0.4.0
mat73
python-dotenv
facenet-pytorch
